requirements.txt
streamlit==1.45.0
google-genai==1.15.0
python-dotenv==1.1.0
